pyproject.toml
[project]
name = "power-source-dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
     "ipykernel>=7.1.0",
     "numpy>=2.3.4",
     "pandas>=2.3.3",
     "plotly>=6.5.2",
     "pyyaml>=6.0.3",
     "streamlit>=1.53.1",
]

[tool.uv.sources]
microscopemetrics = { git = "https://github.com/juliomateoslangerak/microscope-metrics.git", rev = "dev" }
